+++ example/spec/parallel/createReconciliationReport/CreateReconciliationReportSpec.js
@@ -8,12 +8,10 @@ const got = require('got');
 const isEqual = require('lodash/isEqual');
 const isNil = require('lodash/isNil');
 const pWaitFor = require('p-wait-for');
-
 const { deleteAsyncOperation } = require('@cumulus/api-client/asyncOperations');
 const reconciliationReportsApi = require('@cumulus/api-client/reconciliationReports');
-const { deleteExecution } = require('@cumulus/api-client/executions');
 const {
-  buildS3Uri, fileExists, getJsonS3Object, parseS3Uri, s3PutObject, deleteS3Object,
+  buildS3Uri, fileExists, getJsonS3Object, parseS3Uri, s3PutObject,
 } = require('@cumulus/aws-client/S3');
 const { CMR } = require('@cumulus/cmr-client');
 const { lambda, s3 } = require('@cumulus/aws-client/services');
@@ -28,8 +26,6 @@ const { Granule } = require('@cumulus/api/models');
 const {
   addCollections,
   addProviders,
-  buildAndExecuteWorkflow,
-  cleanupCollections,
   cleanupProviders,
   generateCmrXml,
   waitForAsyncOperationStatus,
@@ -38,11 +34,11 @@ const {
 const { getGranuleWithStatus } = require('@cumulus/integration-tests/Granules');
 const { createCollection } = require('@cumulus/integration-tests/Collections');
 const { createProvider } = require('@cumulus/integration-tests/Providers');
-const { deleteCollection, getCollections } = require('@cumulus/api-client/collections');
-const { deleteGranule, getGranule, removePublishedGranule } = require('@cumulus/api-client/granules');
-const { deleteProvider } = require('@cumulus/api-client/providers');
+const { getCollections } = require('@cumulus/api-client/collections');
+const { getGranule } = require('@cumulus/api-client/granules');
 const { getCmrSettings } = require('@cumulus/cmrjs/cmr-utils');
 
+const { buildAndExecuteWorkflow } = require('../../helpers/workflowUtils');
 const {
   loadConfig,
   uploadTestDataToBucket,
@@ -51,11 +47,11 @@ const {
   createTestDataPath,
   createTestSuffix,
 } = require('../../helpers/testUtils');
+const { removeCollectionAndAllDependencies } = require('../../helpers/Collections');
 const {
   setupTestGranuleForIngest, waitForGranuleRecordUpdatedInList,
 } = require('../../helpers/granuleUtils');
 const { waitForModelStatus } = require('../../helpers/apiUtils');
-
 const providersDir = './data/providers/s3/';
 const collectionsDir = './data/collections/s3_MYD13Q1_006';
 const collection = { name: 'MYD13Q1', version: '006' };
@@ -80,7 +76,7 @@ async function setupCollectionAndTestData(config, testSuffix, testDataFolder) {
     '@cumulus/test-data/granules/BROWSE.MYD13Q1.A2002185.h00v09.006.2015149071135.hdf',
     '@cumulus/test-data/granules/BROWSE.MYD13Q1.A2002185.h00v09.006.2015149071135.1.jpg',
   ];
-
+  await removeCollectionAndAllDependencies({ prefix: config.stackName, collection });
   // populate collections, providers and test data
   await Promise.all([
     uploadTestDataToBucket(config.bucket, s3data, testDataFolder),
@@ -92,12 +88,13 @@ async function setupCollectionAndTestData(config, testSuffix, testDataFolder) {
 let ingestGranuleExecutionArn;
 const ingestAndPublishGranuleExecutionArns = [];
 
+// TODO: [CUMULUS-2567] These should be in a helper, possibly unit tested.
 /**
  * Creates a new test collection with associated granule for testing.
  *
  * @param {string} prefix - stack Prefix
  * @param {string} sourceBucket - testing source bucket
- * @returns {Promise<Array>} A new collection with associated granule and a cleanup function to call after you are finished.
+ * @returns {Promise<Object>}  The collection created
  */
 const createActiveCollection = async (prefix, sourceBucket) => {
   // The S3 path where granules will be ingested from
@@ -162,23 +159,7 @@ const createActiveCollection = async (prefix, sourceBucket) => {
   });
 
   await getGranuleWithStatus({ prefix, granuleId, status: 'completed' });
-
-  const cleanupFunction = async () => {
-    await deleteGranule({ prefix, granuleId });
-    await Promise.allSettled(
-      [
-        deleteS3Object(sourceBucket, granFileKey),
-        deleteProvider({ prefix, providerId: get(provider, 'id') }),
-        deleteCollection({
-          prefix,
-          collectionName: get(newCollection, 'name'),
-          collectionVersion: get(newCollection, 'version'),
-        }),
-      ]
-    );
-  };
-
-  return [newCollection, cleanupFunction];
+  return newCollection;
 };
 
 // ingest a granule and publish if requested
@@ -199,9 +180,11 @@ async function ingestAndPublishGranule(config, testSuffix, testDataFolder, publi
     testDataFolder
   );
 
-  ingestAndPublishGranuleExecutionArns.push(await buildAndExecuteWorkflow(
+  const { executionArn } = await buildAndExecuteWorkflow(
     config.stackName, config.bucket, workflowName, collection, provider, inputPayload
-  ));
+  );
+
+  ingestAndPublishGranuleExecutionArns.push(executionArn);
 
   await waitForModelStatus(
     new Granule(),
@@ -301,7 +284,6 @@ describe('When there are granule differences and granule reconciliation is run',
   let config;
   let dbGranuleId;
   let extraCumulusCollection;
-  let extraCumulusCollectionCleanup;
   let extraFileInDb;
   let extraS3Object;
   let granuleBeforeUpdate;
@@ -344,7 +326,7 @@ describe('When there are granule differences and granule reconciliation is run',
       process.env.FilesTable = `${config.stackName}-FilesTable`;
       await GranuleFilesCache.put(extraFileInDb);
 
-      const activeCollectionPromise = createActiveCollection(config.stackName, config.bucket);
+      extraCumulusCollection = await createActiveCollection(config.stackName, config.bucket);
 
       const testId = createTimestampedTestId(config.stackName, 'CreateReconciliationReport');
       testSuffix = createTestSuffix(testId);
@@ -358,12 +340,10 @@ describe('When there are granule differences and granule reconciliation is run',
         publishedGranuleId,
         dbGranuleId,
         cmrGranule,
-        [extraCumulusCollection, extraCumulusCollectionCleanup],
       ] = await Promise.all([
         ingestAndPublishGranule(config, testSuffix, testDataFolder),
         ingestAndPublishGranule(config, testSuffix, testDataFolder, false),
         ingestGranuleToCMR(cmrClient),
-        activeCollectionPromise,
       ]);
 
       console.log('XXXXX Waiting for collections in list');
@@ -397,13 +377,12 @@ describe('When there are granule differences and granule reconciliation is run',
       ]);
     } catch (error) {
       console.log(error);
-      beforeAllFailed = true;
-      throw error;
+      beforeAllFailed = error;
     }
   });
 
   it('prepares the test suite successfully', () => {
-    if (beforeAllFailed) fail('beforeAll() failed to prepare test suite');
+    if (beforeAllFailed) fail(beforeAllFailed);
   });
 
   describe('Create an Inventory Reconciliation Report to monitor inventory discrepancies', () => {
@@ -419,6 +398,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates an async operation through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const response = await reconciliationReportsApi.createReconciliationReport({
         prefix: config.stackName,
         request: {
@@ -437,6 +417,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       let asyncOperation;
       try {
         asyncOperation = await waitForAsyncOperationStatus({
@@ -444,7 +425,7 @@ describe('When there are granule differences and granule reconciliation is run',
           status: 'SUCCEEDED',
           stackName: config.stackName,
           retryOptions: {
-            retries: 70,
+            retries: 80,
             factor: 1.041,
           },
         });
@@ -458,6 +439,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('fetches a reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const reportContent = await fetchReconciliationReport(config.stackName, reportRecord.name);
       report = JSON.parse(reportContent);
       expect(report.reportType).toBe('Granule Not Found');
@@ -465,28 +447,33 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates a report showing cumulus files that are in S3 but not in the DynamoDB Files table', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const extraS3ObjectUri = buildS3Uri(extraS3Object.Bucket, extraS3Object.Key);
       expect(report.filesInCumulus.onlyInS3).toContain(extraS3ObjectUri);
     });
 
     it('generates a report showing cumulus files that are in the DynamoDB Files table but not in S3', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const extraFileUri = buildS3Uri(extraFileInDb.bucket, extraFileInDb.key);
       const extraDbUris = report.filesInCumulus.onlyInDynamoDb.map((i) => i.uri);
       expect(extraDbUris).toContain(extraFileUri);
     });
 
     it('generates a report showing number of collections that are in both Cumulus and CMR', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       // MYD13Q1___006 is in both Cumulus and CMR
       expect(report.collectionsInCumulusCmr.okCount).toBeGreaterThanOrEqual(1);
     });
 
     it('generates a report showing collections that are in Cumulus but not in CMR', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const extraCollection = constructCollectionId(extraCumulusCollection.name, extraCumulusCollection.version);
       expect(report.collectionsInCumulusCmr.onlyInCumulus).toContain(extraCollection);
       expect(report.collectionsInCumulusCmr.onlyInCumulus).not.toContain(collectionId);
     });
 
     it('generates a report showing the amount of files that match broken down by Granule', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const okCount = report.filesInCumulus.okCount;
       const totalOkCountByGranule = Object.values(report.filesInCumulus.okCountByGranule).reduce(
         (total, currentOkCount) => total + currentOkCount
@@ -495,17 +482,20 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates a report showing collections that are in the CMR but not in Cumulus', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       // we know CMR has collections which are not in Cumulus
       expect(report.collectionsInCumulusCmr.onlyInCmr.length).toBe(1);
       expect(report.collectionsInCumulusCmr.onlyInCmr).not.toContain(collectionId);
     });
 
     it('generates a report showing number of granules that are in both Cumulus and CMR', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       // published granule should in both Cumulus and CMR
       expect(report.granulesInCumulusCmr.okCount).toBeGreaterThanOrEqual(1);
     });
 
     it('generates a report showing granules that are in the Cumulus but not in CMR', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       // ingested (not published) granule should only in Cumulus
       const cumulusGranuleIds = report.granulesInCumulusCmr.onlyInCumulus.map((gran) => gran.granuleId);
       expect(cumulusGranuleIds).toContain(dbGranuleId);
@@ -513,6 +503,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates a report showing granules that are in the CMR but not in Cumulus', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const cmrGranuleIds = report.granulesInCumulusCmr.onlyInCmr.map((gran) => gran.GranuleUR);
       expect(cmrGranuleIds.length).toBeGreaterThanOrEqual(1);
       expect(cmrGranuleIds).toContain(cmrGranule.granuleId);
@@ -521,11 +512,13 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates a report showing number of granule files that are in both Cumulus and CMR', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       // published granule should have 2 files in both Cumulus and CMR
       expect(report.filesInCumulusCmr.okCount).toBeGreaterThanOrEqual(2);
     });
 
     it('generates a report showing granule files that are in Cumulus but not in CMR', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       // published granule should have one file(renamed file) in Cumulus
       const fileNames = report.filesInCumulusCmr.onlyInCumulus.map((file) => file.fileName);
       expect(fileNames).toContain(updatedGranuleFile.fileName);
@@ -535,6 +528,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates a report showing granule files that are in the CMR but not in Cumulus', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const urls = report.filesInCumulusCmr.onlyInCmr;
       expect(urls.find((url) => url.URL.endsWith(originalGranuleFile.fileName))).toBeTruthy();
       expect(urls.find((url) => url.URL.endsWith(updatedGranuleFile.fileName))).toBeFalsy();
@@ -544,6 +538,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('deletes a reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       await reconciliationReportsApi.deleteReconciliationReport({
         prefix: config.stackName,
         name: reportRecord.name,
@@ -553,13 +548,19 @@ describe('When there are granule differences and granule reconciliation is run',
       const exists = await fileExists(parsed.Bucket, parsed.Key);
       expect(exists).toBeFalse();
 
-      const response = await reconciliationReportsApi.getReconciliationReport({
-        prefix: config.stackName,
-        name: reportRecord.name,
-      });
+      let responseError;
 
-      expect(response.statusCode).toBe(404);
-      expect(JSON.parse(response.body).message).toBe(`No record found for ${reportRecord.name}`);
+      try {
+        await reconciliationReportsApi.getReconciliationReport({
+          prefix: config.stackName,
+          name: reportRecord.name,
+        });
+      } catch (error) {
+        responseError = error;
+      }
+
+      expect(responseError.statusCode).toBe(404);
+      expect(JSON.parse(responseError.apiMessage).message).toBe(`No record found for ${reportRecord.name}`);
     });
   });
 
@@ -576,6 +577,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates an async operation through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const request = {
         reportType: 'Internal',
         reportName: randomId('InternalReport'),
@@ -595,6 +597,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       let asyncOperation;
       try {
         asyncOperation = await waitForAsyncOperationStatus({
@@ -613,6 +616,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('fetches a reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const reportContent = await fetchReconciliationReport(config.stackName, reportRecord.name);
       report = JSON.parse(reportContent);
       expect(report.reportType).toBe('Internal');
@@ -620,6 +624,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates a report showing number of collections that are in both ES and DB', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       expect(report.collections.okCount).toBe(1);
       expect(report.collections.withConflicts.length).toBe(0);
       expect(report.collections.onlyInEs.length).toBe(0);
@@ -627,6 +632,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates a report showing number of granules that are in both ES and DB', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       expect(report.granules.okCount).toBe(2);
       expect(report.granules.withConflicts.length).toBe(0);
       if (report.granules.withConflicts.length !== 0) {
@@ -637,6 +643,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('deletes a reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       await reconciliationReportsApi.deleteReconciliationReport({
         prefix: config.stackName,
         name: reportRecord.name,
@@ -646,13 +653,18 @@ describe('When there are granule differences and granule reconciliation is run',
       const exists = await fileExists(parsed.Bucket, parsed.Key);
       expect(exists).toBeFalse();
 
-      const response = await reconciliationReportsApi.getReconciliationReport({
-        prefix: config.stackName,
-        name: reportRecord.name,
-      });
+      let responseError;
+      try {
+        await reconciliationReportsApi.getReconciliationReport({
+          prefix: config.stackName,
+          name: reportRecord.name,
+        });
+      } catch (error) {
+        responseError = error;
+      }
 
-      expect(response.statusCode).toBe(404);
-      expect(JSON.parse(response.body).message).toBe(`No record found for ${reportRecord.name}`);
+      expect(responseError.statusCode).toBe(404);
+      expect(JSON.parse(responseError.apiMessage).message).toBe(`No record found for ${reportRecord.name}`);
     });
   });
 
@@ -668,6 +680,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates an async operation through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       const request = {
         reportType: 'Granule Inventory',
         reportName: randomId('granuleInventory'),
@@ -688,6 +701,8 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('generates reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
+
       let asyncOperation;
       try {
         asyncOperation = await waitForAsyncOperationStatus({
@@ -707,9 +722,10 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('Fetches an object with a signedURL to the Granule Inventory report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
+
       const reportContent = await fetchReconciliationReport(config.stackName, reportRecord.name);
       reportArray = reportContent.split('\n');
-
       [
         'granuleUr',
         'collectionId',
@@ -723,6 +739,8 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('includes correct records', () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
+
       [
         collectionId,
         dbGranuleId,
@@ -737,6 +755,7 @@ describe('When there are granule differences and granule reconciliation is run',
     });
 
     it('deletes a reconciliation report through the Cumulus API', async () => {
+      if (beforeAllFailed) fail(beforeAllFailed);
       await reconciliationReportsApi.deleteReconciliationReport({
         prefix: config.stackName,
         name: reportRecord.name,
@@ -746,36 +765,38 @@ describe('When there are granule differences and granule reconciliation is run',
       const exists = await fileExists(parsed.Bucket, parsed.Key);
       expect(exists).toBeFalse();
 
-      const response = await reconciliationReportsApi.getReconciliationReport({
-        prefix: config.stackName,
-        name: reportRecord.name,
-      });
-
-      expect(response.statusCode).toBe(404);
-      expect(JSON.parse(response.body).message).toBe(`No record found for ${reportRecord.name}`);
+      let responseError;
+      try {
+        await reconciliationReportsApi.getReconciliationReport({
+          prefix: config.stackName,
+          name: reportRecord.name,
+        });
+      } catch (error) {
+        responseError = error;
+      }
+      expect(responseError.statusCode).toBe(404);
+      expect(JSON.parse(responseError.apiMessage).message).toBe(`No record found for ${reportRecord.name}`);
     });
   });
 
   afterAll(async () => {
-    console.log(`update granule files back ${publishedGranuleId}`);
+    const activeCollectionId = constructCollectionId(extraCumulusCollection.name, extraCumulusCollection.version);
+    console.log(`update database state back for  ${publishedGranuleId}, ${activeCollectionId}`);
     await granuleModel.update({ granuleId: publishedGranuleId }, { files: granuleBeforeUpdate.files });
-    await removePublishedGranule({
-      prefix: config.stackName,
-      granuleId: publishedGranuleId,
-    });
-
-    // TODO there may be another execution to delete here
-    await deleteExecution({ prefix: config.stackName, executionArn: ingestGranuleExecutionArn });
-    await Promise.all(ingestAndPublishGranuleExecutionArns.map((executionArn) => deleteExecution({ prefix: config.stackName, executionArn })));
-
-    await Promise.all([
+    const cleanupResults = await Promise.allSettled([
+      removeCollectionAndAllDependencies({ prefix: config.stackName, collection: extraCumulusCollection }),
+      removeCollectionAndAllDependencies({ prefix: config.stackName, collection: collection }),
       s3().deleteObject(extraS3Object).promise(),
       GranuleFilesCache.del(extraFileInDb),
       deleteFolder(config.bucket, testDataFolder),
-      cleanupCollections(config.stackName, config.bucket, collectionsDir),
-      cleanupProviders(config.stackName, config.bucket, providersDir, testSuffix),
-      extraCumulusCollectionCleanup(),
       cmrClient.deleteGranule(cmrGranule),
     ]);
+    cleanupResults.forEach((result) => {
+      if (result.status === 'rejected') {
+        console.log(`***Cleanup failed ${JSON.stringify(result)}`);
+        throw new Error(JSON.stringify(result));
+      }
+    });
+    await cleanupProviders(config.stackName, config.bucket, providersDir, testSuffix);
   });
 });